The responsiveness of the entire site was changed and improved. The reorganization of some files was also included, such as: "thanks", "voluntary" and removal of "creditUser" and "creditVolunteer".
