+++ src/components/Header/header.css
@@ -52,3 +52,54 @@
 .logout:hover {
     cursor: pointer;
 }
+
+/* Adicione isso ao final do arquivo */
+
+@media (max-width: 768px) {
+    .headerContent {
+        box-sizing: border-box;
+        align-items: center;
+        padding: 10px;
+    }
+
+    .headerLogo img {
+        width: 50px;
+        height: auto;
+    }
+
+    .toolsBar nav {
+        display: flex;
+        flex-direction: row;
+        align-items: center;
+        gap: 15px;
+    }
+
+    .logout {
+        margin-top: 10px;
+        margin-right: 0;
+    }
+}
+
+@media (max-width: 480px) {
+    .headerLine {
+        padding: 10px;
+        box-sizing: border-box;
+        position: absolute;
+    }
+
+    .headerLogo img {
+        width: 40px;
+    }
+
+    .toolsBar ul {
+        gap: 15px;
+    }
+
+    .toolsBar li img {
+        width: 24px;
+    }
+
+    .logout img {
+        width: 24px;
+    }
+}
