+++ src/components/Home/homepage.css
@@ -6,6 +6,11 @@
 .homepage {
     width: 100%;
     height: 100%;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    padding-top: 50px;
+    /* Adicionando padding-top para evitar sobreposição com o header */
 }
 
 .no-posts-message p {
@@ -13,7 +18,7 @@
     font-size: 20px;
     font-weight: 400;
     text-align: center;
-    Color: #000000A6;
+    color: #000000A6;
 }
 
 .no-posts-message h1 {
@@ -22,4 +27,37 @@
     font-weight: 400;
     text-align: center;
     color: #000000A6;
+}
+
+/* Adicione isso ao final do arquivo */
+
+@media (max-width: 768px) {
+    .homepage {
+        flex-direction: column;
+        padding: 60px 20px 20px 20px;
+        /* Ajustando o padding-top para telas menores */
+    }
+
+    .no-posts-message p {
+        font-size: 18px;
+    }
+
+    .no-posts-message h1 {
+        font-size: 22px;
+    }
+}
+
+@media (max-width: 480px) {
+    .homepage {
+        padding: 60px 10px 10px 10px;
+        box-sizing: border-box;
+    }
+
+    .no-posts-message p {
+        font-size: 16px;
+    }
+
+    .no-posts-message h1 {
+        font-size: 20px;
+    }
 }
