+++ src/components/UserPost/Post.css
@@ -2,7 +2,6 @@
     width: 100%;
     height: 100%;
     display: flex;
-    flex-direction: column;
     align-items: center;
 }
