--- src/components/cadastro/voluntária/formulario/formulario.css
@@ -1,36 +1,0 @@
-.questionary-page {
-    background: #F2E1E9;
-    display: flex;
-    justify-content: center;
-    height: 100svh;
-}
-
-.questionary-page h1 {
-    font-family: Yrsa;
-    font-size: 27px;
-    font-weight: 400;
-    text-align: center;
-}
-
-.questionary-page hr {
-    width: 100%;
-    top: 90px;
-    left: 14px;
-    border: 1px 0px 0px 0px;
-    color: rgba(0, 0, 0, 0.568);
-}
-
-.questionary-box {
-    width: 629px;
-}
-
-.questionary-selection p {
-    font-family: Tenor Sans;
-    font-size: 20px;
-    font-weight: 400;
-    line-height: 17.55px;
-    text-align: center;
-
-}
-
-
