--- src/components/cadastro/voluntária/voluntaria.css
@@ -1,121 +1,0 @@
-.volunteering {
-    display: flex;
-    align-items: center;
-    flex-direction: column;
-    background: #eebbd38a;
-    height: 100vh;
-
-}
-
-.volunteering-description {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    margin: 15px;
-    width: 80%;
-}
-
-.volunteering-description h1 {
-    font-family: Yrsa;
-    font-size: 30px;
-    font-weight: 300;
-    text-align: center;
-}
-
-.volunteering-description h2 {
-    font-family: Yrsa;
-    font-size: 25px;
-    font-weight: 400;
-    text-align: center;
-}
-
-.volunteering-cards {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    gap: 20px;
-    margin-bottom: 20px;
-}
-
-.volunteering-cards-accepted {
-    background: #EEBBD3;
-    display: flex;
-    align-items: center;
-    padding: 15px;
-    border-radius: 20px;
-    width: 330px;
-
-}
-
-.volunteering-cards-accepted img {
-    Width: 80px;
-    Height: 80px;
-    background-image: "../../../../img/icones/voluntario/voluntaria.png";
-}
-
-.volunteering-cards-accepted a {
-    font-family: Yrsa;
-    font-size: 25px;
-    font-weight: 400;
-    text-align: center;
-    font-style: normal;
-    color: black;
-    text-decoration: none;
-    position: relative;
-}
-
-.volunteering-cards-accepted a::after {
-    content: '';
-    position: absolute;
-    left: 0;
-    bottom: -3px; 
-    width: 0; 
-    height: 2px; 
-    background-color: rgba(0, 0, 0, 0.466); 
-    transition: width 0.3s ease; 
-}
-
-.volunteering-cards-accepted a:hover::after {
-    width: 100%; 
-}
-
-.volunteering-cards-rejected {
-    background: #EEBBD3;
-    display: flex;
-    align-items: center;
-    padding: 15px;
-    border-radius: 20px;
-    width: 330px;
-}
-
-.volunteering-cards-rejected img {
-    Width: 80px;
-    Height: 80px;
-    background-image: "../../../../img/icones/voluntario/usuaria.png";
-}
-
-.volunteering-cards-rejected a {
-    font-family: Yrsa;
-    font-size: 25px;
-    font-weight: 400;
-    text-align: center;
-    font-style: normal;
-    color: black;
-    text-decoration: none;
-    position: relative;
-}
-
-.volunteering-cards-rejected a::after {
-    content: '';
-    position: absolute;
-    left: 0;
-    bottom: -3px; 
-    width: 0; 
-    height: 2px; 
-    background-color: rgba(0, 0, 0, 0.466); 
-    transition: width 0.3s ease; 
-}
-
-.volunteering-cards-rejected a:hover::after {
-    width: 100%; 
-}
